# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/articles": {
      "get": {
        "description": "前台首页与各筛选页共用的接口。只返回 `status=published` 且**发布时间已到**的文章，\n所以「定时发布」的文章在到点前不会出现在这里，无需任何定时任务。\n\n**关键词搜索的实现说明**：标题与摘要走 `pg_trgm` 三元组索引（`ILIKE`，对中文有效），\n正文走 `tsvector`（对英文有效），两者取并集后按 `ts_rank` 排序。",
        "operationId": "ArticlesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 12,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词。标题/摘要走 pg_trgm 三元组索引（对中文有效），正文走 tsvector（对英文有效）",
            "schema": {
              "example": "数据库",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "description": "按分类 slug 筛选",
            "schema": {
              "example": "backend",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "按分类 ID 筛选",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "tagSlug",
            "required": false,
            "in": "query",
            "description": "按标签 slug 筛选",
            "schema": {
              "example": "postgresql",
              "type": "string"
            }
          },
          {
            "name": "authorUsername",
            "required": false,
            "in": "query",
            "description": "按作者用户名筛选",
            "schema": {
              "example": "zhangsan",
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "按年份筛选（归档页用）",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "按月份筛选（归档页用，1–12）",
            "schema": {
              "example": 9,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "排序：latest=最新发布 / oldest=最早 / hot=按阅读量 / comments=按评论数",
            "schema": {
              "default": "latest",
              "type": "string",
              "enum": [
                "latest",
                "oldest",
                "hot",
                "comments"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页文章列表"
          }
        },
        "summary": "文章列表（公开）",
        "tags": [
          "文章 · 前台"
        ]
      }
    },
    "/api/v1/articles/archives": {
      "get": {
        "description": "归档页数据。年月聚合在数据库里完成（`EXTRACT` + `GROUP BY`），不把全表拉到内存。",
        "operationId": "ArticlesController_archives",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 12,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词。标题/摘要走 pg_trgm 三元组索引（对中文有效），正文走 tsvector（对英文有效）",
            "schema": {
              "example": "数据库",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "description": "按分类 slug 筛选",
            "schema": {
              "example": "backend",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "按分类 ID 筛选",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "tagSlug",
            "required": false,
            "in": "query",
            "description": "按标签 slug 筛选",
            "schema": {
              "example": "postgresql",
              "type": "string"
            }
          },
          {
            "name": "authorUsername",
            "required": false,
            "in": "query",
            "description": "按作者用户名筛选",
            "schema": {
              "example": "zhangsan",
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "按年份筛选（归档页用）",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "按月份筛选（归档页用，1–12）",
            "schema": {
              "example": 9,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "排序：latest=最新发布 / oldest=最早 / hot=按阅读量 / comments=按评论数",
            "schema": {
              "default": "latest",
              "type": "string",
              "enum": [
                "latest",
                "oldest",
                "hot",
                "comments"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "归档分组列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArchiveItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "按年月归档",
        "tags": [
          "文章 · 前台"
        ]
      }
    },
    "/api/v1/articles/{slug}": {
      "get": {
        "description": "支持**可选认证**：带有效 Token 时会在响应里回填 `likedByMe`（配合 IP 判断）与 `canEdit`，\n前端无需再发一次「我是谁」的请求。\n\n草稿与未到发布时间的文章只有作者本人和管理员可见，其他人一律得到 404（不是 403 —— \n用 403 会暴露「这个 slug 确实存在」这一信息）。",
        "operationId": "ArticlesController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "文章 URL 标识",
            "schema": {
              "example": "postgresql-中文全文检索",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文章详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "文章不存在或尚未发布",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "文章详情（公开）",
        "tags": [
          "文章 · 前台"
        ]
      }
    },
    "/api/v1/articles/{id}/view": {
      "post": {
        "description": "同 IP 对同一篇文章 10 分钟内只计一次，防止刷新刷量。重复上报返回当前值且不报错（幂等）。",
        "operationId": "ArticlesController_view",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "当前累计阅读量",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewResultDto"
                }
              }
            }
          }
        },
        "summary": "上报阅读量",
        "tags": [
          "文章 · 前台"
        ]
      }
    },
    "/api/v1/articles/{id}/like": {
      "post": {
        "description": "按 **IP + 文章** 去重（`article_likes` 表的唯一约束）。重复点赞是幂等的，返回 `changed=false`。",
        "operationId": "ArticlesController_like",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "点赞结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeResultDto"
                }
              }
            }
          }
        },
        "summary": "点赞",
        "tags": [
          "文章 · 前台"
        ]
      },
      "delete": {
        "description": "取消当前 IP 的点赞记录。未点过赞时返回 `changed=false`，不报错。",
        "operationId": "ArticlesController_unlike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeResultDto"
                }
              }
            }
          }
        },
        "summary": "取消点赞",
        "tags": [
          "文章 · 前台"
        ]
      }
    },
...
    "/api/v1/me/articles/{id}": {
      "get": {
        "description": "草稿也可以取到，仅限作者本人与管理员。",
        "operationId": "MeArticlesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "example": "10",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文章管理视图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleManageDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "不是自己的文章",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "取文章原始内容（编辑器回填）",
        "tags": [
          "文章 · 作者工作台"
        ]
      },
      "patch": {
        "description": "仅作者本人或管理员。正文变更会同步重算 HTML 缓存与摘要。",
        "operationId": "MeArticlesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleManageDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "无权操作他人的文章",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新文章",
        "tags": [
          "文章 · 作者工作台"
        ]
      },
      "delete": {
        "description": "软删除，数据仍在库里，`status` 变为 `deleted` 并记录 `deletedAt`，可随时恢复。",
        "operationId": "MeArticlesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已移入回收站",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResultDto"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除文章（移入回收站）",
        "tags": [
          "文章 · 作者工作台"
        ]
      }
    },
...
  },
  "info": {
    "title": "博客平台 API",
    "description": "多作者博客平台后端接口。\n\n## 认证方式\n登录后拿到 `accessToken`，在请求头带上 `Authorization: Bearer <token>`。\nAccess Token 默认 15 分钟过期，用 `POST /auth/refresh` 换成新的令牌对。\n\n## 关于 ID\n所有主键在数据库里是 `BIGINT`，接口统一以**字符串**收发 ——\nJS 的 `Number` 在 2^53 以上会静默丢精度，这是比报错更难排查的问题。\n\n## 关于分页\n列表接口统一返回 `{ items, total, page, pageSize, totalPages }`。\n\n## 关于权限\n角色两级：`admin`（全站）与 `author`（只能操作自己的内容）。\n另外账号需要处于 `active` 状态才能执行写操作 —— 注册后是 `pending`，需管理员审核激活。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "填入 accessToken（不含 Bearer 前缀）"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP 状态码",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "错误简述",
            "example": "Bad Request"
          },
          "message": {
            "description": "错误详情；校验失败时为字符串数组",
            "example": [
              "password 长度不能少于 8 位"
            ],
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/v1/articles"
          },
          "timestamp": {
            "type": "string",
            "description": "发生时间（ISO 8601）",
            "example": "2026-09-17T09:30:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "OperationResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "附带说明",
            "example": "已移入回收站"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "ArticleAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "作者 ID",
            "example": "1"
          },
          "username": {
            "type": "string",
            "description": "用户名，可跳转 /author/{username}",
            "example": "zhangsan"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "张三"
          },
          "avatarUrl": {
            "type": "string",
            "description": "头像",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "nickname"
        ]
      },
      "ArticleCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "分类 ID",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "分类名",
            "example": "后端"
          },
          "slug": {
            "type": "string",
            "description": "分类 slug",
            "example": "backend"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
...
      "ArticleListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "文章 ID",
            "example": "10"
          },
          "title": {
            "type": "string",
            "description": "标题",
            "example": "PostgreSQL 中文全文检索的正确姿势"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识",
            "example": "postgresql-中文全文检索"
          },
          "summary": {
            "type": "string",
            "description": "摘要",
            "example": "tsvector 的 simple 配置对中文不分词……"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图",
            "nullable": true
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶",
            "example": false
          },
          "viewCount": {
            "type": "number",
            "description": "阅读量",
            "example": 128
          },
          "likeCount": {
            "type": "number",
            "description": "点赞数",
            "example": 12
          },
          "commentCount": {
            "type": "number",
            "description": "已通过的评论数",
            "example": 5
          },
          "publishedAt": {
            "type": "string",
            "description": "发布时间；为 null 表示尚未发布",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "description": "最后更新时间"
          },
          "author": {
            "description": "作者",
            "allOf": [
              {
                "$ref": "#/components/schemas/ArticleAuthorDto"
              }
            ]
          },
          "category": {
            "description": "分类；未分类时为 null",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ArticleCategoryDto"
              }
            ]
          },
          "tags": {
            "description": "标签列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTagDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "isTop",
          "viewCount",
          "likeCount",
          "commentCount",
          "updatedAt",
          "author",
          "tags"
        ]
      },
...
      "ArchiveItemDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "description": "年份",
            "example": 2026
          },
          "month": {
            "type": "number",
            "description": "月份",
            "example": 9
          },
          "count": {
            "type": "number",
            "description": "该月文章数",
            "example": 7
          },
          "articles": {
            "description": "该月文章（按时间倒序，最多 50 篇）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleListItemDto"
            }
          }
        },
        "required": [
          "year",
          "month",
          "count",
          "articles"
        ]
      },
...
      "ArticleDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "文章 ID",
            "example": "10"
          },
          "title": {
            "type": "string",
            "description": "标题",
            "example": "PostgreSQL 中文全文检索的正确姿势"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识",
            "example": "postgresql-中文全文检索"
          },
          "summary": {
            "type": "string",
            "description": "摘要",
            "example": "tsvector 的 simple 配置对中文不分词……"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图",
            "nullable": true
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶",
            "example": false
          },
          "viewCount": {
            "type": "number",
            "description": "阅读量",
            "example": 128
          },
          "likeCount": {
            "type": "number",
            "description": "点赞数",
            "example": 12
          },
          "commentCount": {
            "type": "number",
            "description": "已通过的评论数",
            "example": 5
          },
          "publishedAt": {
            "type": "string",
            "description": "发布时间；为 null 表示尚未发布",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "description": "最后更新时间"
          },
          "author": {
            "description": "作者",
            "allOf": [
              {
                "$ref": "#/components/schemas/ArticleAuthorDto"
              }
            ]
          },
          "category": {
            "description": "分类；未分类时为 null",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ArticleCategoryDto"
              }
            ]
          },
          "tags": {
            "description": "标签列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTagDto"
            }
          },
          "content": {
            "type": "string",
            "description": "Markdown 原文（编辑器回填用）"
          },
          "contentHtml": {
            "type": "string",
            "description": "渲染并净化后的 HTML，可直接 v-html 注入"
          },
          "toc": {
            "description": "正文目录（h2–h4）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TocItemDto"
            }
          },
          "authorBio": {
            "type": "string",
            "description": "作者简介",
            "nullable": true
          },
          "authorGithub": {
            "type": "string",
            "description": "作者 GitHub",
            "nullable": true
          },
          "likedByMe": {
            "type": "boolean",
            "description": "当前请求方是否已点赞过（按 IP + 文章去重）",
            "example": false
          },
          "canEdit": {
            "type": "boolean",
            "description": "当前请求方是否可以编辑这篇文章（作者本人或管理员）",
            "example": false
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "isTop",
          "viewCount",
          "likeCount",
          "commentCount",
          "updatedAt",
          "author",
          "tags",
          "content",
          "contentHtml",
          "toc",
          "authorBio",
          "authorGithub",
          "likedByMe",
          "canEdit"
        ]
      },
      "ViewResultDto": {
        "type": "object",
        "properties": {
          "viewCount": {
            "type": "number",
            "description": "上报后该文章的累计阅读量",
            "example": 129
          }
        },
        "required": [
          "viewCount"
        ]
      },
      "LikeResultDto": {
        "type": "object",
        "properties": {
          "likeCount": {
            "type": "number",
            "description": "该文章当前点赞总数",
            "example": 13
          },
          "liked": {
            "type": "boolean",
            "description": "当前请求方是否已点赞",
            "example": true
          },
          "changed": {
            "type": "boolean",
            "description": "本次操作是否改变了状态（重复点赞会被幂等忽略）",
            "example": true
          }
        },
        "required": [
          "likeCount",
          "liked",
          "changed"
        ]
      },
      "ArticleManageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "文章 ID",
            "example": "10"
          },
          "title": {
            "type": "string",
            "description": "标题"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识"
          },
          "content": {
            "type": "string",
            "description": "正文 Markdown 原文"
          },
          "summary": {
            "type": "string",
            "description": "摘要"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "draft",
              "published",
              "deleted"
            ],
            "example": "draft"
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶"
          },
          "categoryId": {
            "type": "string",
            "description": "分类 ID",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "description": "分类名",
            "nullable": true
          },
          "tags": {
            "description": "标签名数组",
            "example": [
              "PostgreSQL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "viewCount": {
            "type": "number",
            "description": "阅读量"
          },
          "likeCount": {
            "type": "number",
            "description": "点赞数"
          },
          "commentCount": {
            "type": "number",
            "description": "评论数"
          },
          "publishedAt": {
            "type": "string",
            "description": "计划/实际发布时间",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "description": "移入回收站的时间",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间"
          },
          "authorId": {
            "type": "string",
            "description": "作者 ID"
          },
          "authorNickname": {
            "type": "string",
            "description": "作者昵称"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "content",
          "summary",
          "status",
          "isTop",
          "tags",
          "viewCount",
          "likeCount",
          "commentCount",
          "createdAt",
          "updatedAt",
          "authorId",
          "authorNickname"
        ]
      },
...
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "标题",
            "example": "PostgreSQL 中文全文检索的正确姿势",
            "maxLength": 200
          },
          "content": {
            "type": "string",
            "description": "正文（Markdown 原文）。服务端会渲染成 HTML 并做 XSS 白名单过滤后缓存",
            "example": "# 标题\n\n正文内容"
          },
          "summary": {
            "type": "string",
            "description": "摘要；留空则自动从正文截取前 200 字"
          },
          "slug": {
            "type": "string",
            "description": "URL 标识；留空则根据标题自动生成（中文保留原字）"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图地址"
          },
          "categoryId": {
            "type": "string",
            "description": "分类 ID（只能选用管理员已创建的分类）",
            "example": "1",
            "nullable": true
          },
          "tags": {
            "description": "标签名数组。不存在的标签会被自动创建（FR-3.2 作者可自由创建标签）",
            "example": [
              "PostgreSQL",
              "全文检索"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "发布状态。draft=存为草稿；published=发布（配合 publishedAt 可做定时发布）",
            "enum": [
              "draft",
              "published"
            ],
            "default": "draft"
          },
          "publishedAt": {
            "type": "string",
            "description": "发布时间（ISO 8601）。传未来时间即为「定时发布」——前台在到点前不会展示，到点自动可见，无需定时任务。不传则取当前时间。",
            "example": "2026-10-01T09:00:00.000Z"
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶（作者可置顶自己的文章）",
            "default": false
          }
        }
      },
...
    }
  }
}